Clean up internal/testutil/helpers package.
