+++ mongo/description/selector_spec_test.go
@@ -10,7 +10,7 @@ import (
 	"path"
 	"testing"
 
-	testhelpers "go.mongodb.org/mongo-driver/internal/testutil/helpers"
+	"go.mongodb.org/mongo-driver/internal/testutil/helpers"
 )
 
 const selectorTestsDir = "../../data/server-selection/server_selection"
@@ -28,7 +28,7 @@ func TestServerSelectionSpec(t *testing.T) {
 		for _, subdir := range [...]string{"read", "write"} {
 			subdirPath := path.Join(topology, subdir)
 
-			for _, file := range testhelpers.FindJSONFilesInDir(t,
+			for _, file := range helpers.FindJSONFilesInDir(t,
 				path.Join(selectorTestsDir, subdirPath)) {
 
 				runTest(t, selectorTestsDir, subdirPath, file)
