+++ mongo/integration/unified/collection_data.go
@@ -12,7 +12,6 @@ import (
 	"fmt"
 
 	"go.mongodb.org/mongo-driver/bson"
-	testhelpers "go.mongodb.org/mongo-driver/internal/testutil/helpers"
 	"go.mongodb.org/mongo-driver/mongo/integration/mtest"
 	"go.mongodb.org/mongo-driver/mongo/options"
 	"go.mongodb.org/mongo-driver/mongo/readconcern"
@@ -71,11 +70,20 @@ func (c *collectionData) createCollection(ctx context.Context) error {
 		return nil
 	}
 
-	docs := testhelpers.RawSliceToInterfaceSlice(c.Documents)
+	docs := rawSliceToInterfaceSlice(c.Documents)
 	if _, err := coll.InsertMany(ctx, docs); err != nil {
 		return fmt.Errorf("error inserting data: %v", err)
 	}
 	return nil
+}
+
+// rawSliceToInterfaceSlice converts a []bson.Raw to []interface{}.
+func rawSliceToInterfaceSlice(elems []bson.Raw) []interface{} {
+	out := make([]interface{}, 0, len(elems))
+	for _, elem := range elems {
+		out = append(out, elem)
+	}
+	return out
 }
 
 // verifyContents asserts that the collection on the server represented by this collectionData instance contains the
