+++ x/mongo/driver/topology/topology_test.go
@@ -21,7 +21,7 @@ import (
 	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/internal"
 	"go.mongodb.org/mongo-driver/internal/testutil/assert"
-	testhelpers "go.mongodb.org/mongo-driver/internal/testutil/helpers"
+	"go.mongodb.org/mongo-driver/internal/testutil/helpers"
 	"go.mongodb.org/mongo-driver/mongo/address"
 	"go.mongodb.org/mongo-driver/mongo/description"
 	"go.mongodb.org/mongo-driver/mongo/readpref"
@@ -675,7 +675,7 @@ type inWindowTestCase struct {
 func TestServerSelectionSpecInWindow(t *testing.T) {
 	const testsDir = "../../../../data/server-selection/in_window"
 
-	files := testhelpers.FindJSONFilesInDir(t, testsDir)
+	files := helpers.FindJSONFilesInDir(t, testsDir)
 
 	for _, file := range files {
 		t.Run(file, func(t *testing.T) {
